sales-incremental-etl/requirements.txt
psycopg2-binary
sqlalchemy
pandas
pyyaml
